feat: quiz UI completed
